Made timing_guard generic, moved experiment-specific things to compare!
Made the timing guard work just like any other metric call, where you
pass a name and a list of tags, and stripped all experiment-specific
code from the guard itself (it no longer adds additional tags, it reports
the metric with whatever name you pass it).

This is to make the guard itself a general-purpose abstraction - if a user
just wants to always time how long something takes, but doesn't want to
muck around with the slightly hard to use time() call (since it requires
you to pass an FnOnce with no return arguments, which is far less
flexible than simply handing out a guard), they have this.

All of the experiment-specific functionality now lives in the compare!
macro which does the tag-shuffling and metric naming for you.

Also added documentation and some tests for the macro. The guard is...
pretty painful to test in any meaningful way, since it necessarily has
to call out to the global Datadog:: fn's. I could have made the enter
function return an Option<TimingGuard>, in which case I could have
potentially given the TimingGuard the ability to hang onto a reference
to the client it needs to call, and therefore allowed for some
dependency injection, but honestly that would have been a bit of a
nightmare both to use and to implement, and the underlying functions the
guard calls are already tested elsewhere, so I am comfortable with the
coverage I have acheived without that